+++ test/integration/tests/4095-utf8-pure-nix/files/ShowUnicode.hs
@@ -1,0 +1,5 @@
+import System.IO (stdout)
+import Text.Printf (hPrintf)
+
+main :: IO ()
+main = hPrintf stdout "平和"